style: cleanup code and removed some comments
