+++ src/types/Account/methods/createTransaction.spec.js
@@ -1,69 +1,18 @@
-const {expect} = require('chai');
-const {HDPrivateKey} = require('@dashevo/dashcore-lib');
+const _ = require('lodash');
+const { expect } = require('chai');
+const { HDPrivateKey } = require('@dashevo/dashcore-lib');
 
 const createTransaction = require('./createTransaction');
-const mnemonic = require('../../../../fixtures/wallets/mnemonics/during-develop-before').mnemonic;
+const { mnemonic } = require('../../../../fixtures/wallets/mnemonics/during-develop-before');
 const FakeNet = require('../../../../fixtures/FakeNet/FakeNet');
 
-const _ = require('lodash');
-// const createTransaction = require('./createTransaction');
 const getUTXOS = require('./getUTXOS');
-const {simpleDescendingAccumulator} = require('../../../utils/coinSelections/strategies');
+const { simpleDescendingAccumulator } = require('../../../utils/coinSelections/strategies');
 
 const addressesFixtures = require('../../../../fixtures/addresses.json');
 const validStore = require('../../../../fixtures/walletStore').valid.orange.store;
 
-const craftedGenerousMinerStrategy = (utxosList, outputsList, deductFee = false, feeCategory = 'normal') => {
-  const TransactionEstimator = require('../../../utils/coinSelections/TransactionEstimator.js');
-  const {sortAndVerifyUTXOS} = require('../../../utils/coinSelections/helpers');
-
-  const txEstimator = new TransactionEstimator(feeCategory);
-
-  // We add our outputs, theses will change only in case deductfee being true
-  txEstimator.addOutputs(outputsList);
-
-  const sort = [{sortBy: 'satoshis', direction: 'descending'}];
-  const sortedUtxosList = sortAndVerifyUTXOS(utxosList, sort);
-
-  const totalOutputValue = txEstimator.getTotalOutputValue();
-
-  let pendingSatoshis = 0;
-  const simplyAccumulatedUtxos = sortedUtxosList.filter((utxo) => {
-    if (pendingSatoshis < totalOutputValue) {
-      pendingSatoshis += utxo.satoshis;
-      return utxo;
-    }
-    return false;
-  });
-  if (pendingSatoshis < totalOutputValue) {
-    throw new Error('Unsufficient utxo amount');
-  }
-
-  // We add the expected inputs, which should match the requested amount
-  // TODO : handle case when we do not match it.
-  txEstimator.addInputs(simplyAccumulatedUtxos);
-
-  const estimatedFee = txEstimator.getFeeEstimate() + 10;
-  if (deductFee === true) {
-    // Then we check that we will be able to do it
-    const inValue = txEstimator.getInValue();
-    const outValue = txEstimator.getOutValue();
-    if (inValue < outValue + estimatedFee) {
-      // We don't have enough change for fee, so we remove from outValue
-      txEstimator.reduceFeeFromOutput((outValue + estimatedFee) - inValue);
-    } else {
-      // TODO : Here we can add some process to check up that we clearly have enough to deduct fee
-    }
-  }
-
-  return {
-    utxos: txEstimator.getInputs(),
-    outputs: txEstimator.getOutputs(),
-    feeCategory,
-    estimatedFee,
-    utxosValue: txEstimator.getInValue(),
-  };
-};
+const craftedGenerousMinerStrategy = require('../../../../fixtures/strategies/craftedGenerousMinerStrategy');
 
 describe('Account - createTransaction', () => {
   let mockWallet;
