+++ src/types/Storage/methods/announce.spec.js
@@ -2,5 +2,5 @@ const { expect } = require('chai');
 const announce = require('./announce');
 
 describe('Storage - announce', () => {
-  // TODO
+  describe.skip('TODO test');
 });
